fix WIR harvesting

+++ harvest_dpird/DPIRD.sh
@@ -3,6 +3,8 @@
 . ./security/dpird.x
 . ./common/start.sh
 
+#export DEBUG=true
+
 FILE=/tmp/tmpx_$$.json
 
 OUTFILE=""
@@ -40,13 +42,18 @@ extract_entry () {
    ATMIN=`echo $AIRTEMP | cut -f2 -d\: | cut -f1 -d,`
    ATMAX=`echo $AIRTEMP | cut -f6 -d\: | cut -f1 -d,`
    ATAVG=`echo $AIRTEMP | cut -f10 -d\: | cut -f1 -d,`
-#  echo airTemperature.min = $ATMIN
-#  echo airTemperature.max = $ATMAX
-#  echo airTemperature.avg = $ATAVG
+if [ "$DEBUG" = "true" ] ; then
+   echo airTemperature.min = $ATMIN
+   echo airTemperature.max = $ATMAX
+   echo airTemperature.avg = $ATAVG
+fi
    read line  # should be '}'
 
    # "relativeHumidity"
    read line  # should be ',"relativeHumidity":'
+if [ "$DEBUG" = "true" ] ; then
+echo RELHUM $line
+fi
    read line  # should be '{'
 
    read RELHUM # should be like :
@@ -54,13 +61,18 @@ extract_entry () {
    RLMIN=`echo $RELHUM | cut -f2 -d\: | cut -f1 -d,`
    RLMAX=`echo $RELHUM | cut -f6 -d\: | cut -f1 -d,`
    RLAVG=`echo $RELHUM | cut -f10 -d\: | cut -f1 -d,`
-#  echo relativeHumidity.min = $RLMIN
-#  echo relativeHumidity.max = $RLMAX
-#  echo relativeHumidity.avg = $RLAVG
+if [ "$DEBUG" = "true" ] ; then
+   echo relativeHumidity.min = $RLMIN
+   echo relativeHumidity.max = $RLMAX
+   echo relativeHumidity.avg = $RLAVG
+fi
    read line  # should be '}'
 
    # "dewPoint"
    read line  # should be ',"dewPoint":'
+if [ "$DEBUG" = "true" ] ; then
+echo DEWPOINT $line
+fi
    read line  # should be '{'
 
    read DEWPOINT # should be like :
@@ -68,57 +80,69 @@ extract_entry () {
    DWMIN=`echo $DEWPOINT | cut -f2 -d\: | cut -f1 -d,`
    DWMAX=`echo $DEWPOINT | cut -f3 -d\: | cut -f1 -d,`
    DWAVG=`echo $DEWPOINT | cut -f4 -d\: | cut -f1 -d,`
-#  echo dewPoint.min = $DWMIN
-#  echo dewPoint.max = $DWMAX
-#  echo dewPoint.avg = $DWAVG
+if [ "$DEBUG" = "true" ] ; then
+   echo dewPoint.min = $DWMIN
+   echo dewPoint.max = $DWMAX
+   echo dewPoint.avg = $DWAVG
+fi
    read line  # should be '}'
 
    # "evapotranspiration"
    read line  # should be ',"evapotranspiration":'
+if [ "$DEBUG" = "true" ] ; then
+echo EVAPO $line
+fi
    read line  # should be '{'
 
    read EVAPTRS # should be like :
    # "shortCrop":0.24,"tallCrop":0.27
    EVSHRT=`echo $EVAPTRS | cut -f2 -d\: | cut -f1 -d,`
    EVTALL=`echo $EVAPTRS | cut -f3 -d\: | cut -f1 -d,`
-#  echo evapotranspiration.shortCrop = $EVSHRT
-#  echo evapotranspiration.tallCrop = $EVTALL
-   read line  # should be '}'
-
-   # "wetBulb"
-   read line  # should be ',"wetBulb":'
-   read line  # should be '{'
-
-   read WETBULB # should be like :
-   # "max":15.7,"avg":15
-   WBMAX=`echo $WETBULB | cut -f2 -d\: | cut -f1 -d,`
-   WBAVG=`echo $WETBULB | cut -f3 -d\:`
-#  echo wetBulb.max = $WBMAX
-#  echo wetBulb.avg = $WBAVG
+if [ "$DEBUG" = "true" ] ; then
+   echo evapotranspiration.shortCrop = $EVSHRT
+   echo evapotranspiration.tallCrop = $EVTALL
+fi
    read line  # should be '}'
 
    # "solarExposure"
    read SOLAR  # should be ',"solarExposure":<value>,"rainfall":<value>,"soilTemperature":'
                # ,"solarExposure":1308.6,"rainfall":0,"soilTemperature":
+if [ "$DEBUG" = "true" ] ; then
+echo SOLAR $SOLAR
+fi
    SOLARTOT=`echo $SOLAR | cut -f2 -d\: | cut -f1 -d,`
    RAINFALL=`echo $SOLAR | cut -f3 -d\: | cut -f1 -d,`
-#  echo solarExposure.solarExposure = $SOLARTOT
-#  echo solarExposure.rainfall = $RAINFALL
+if [ "$DEBUG" = "true" ] ; then
+   echo solarExposure.solarExposure = $SOLARTOT
+   echo solarExposure.rainfall = $RAINFALL
+fi
 
-   read line  # should be '{'
+   SOILTMP=`echo $SOLAR | cut -f4 -d\: | cut -f1 -d,`
+   if [ "$SOILTMP" = "" ] ; then
+      read line  # should be '{'
 
-   read SOILTMP # should be like :
-   # "min":24.3,"max":24.7,"avg":24.5
-   SLMIN=`echo $SOILTMP | cut -f2 -d\: | cut -f1 -d,`
-   SLMAX=`echo $SOILTMP | cut -f3 -d\: | cut -f1 -d,`
-   SLAVG=`echo $SOILTMP | cut -f4 -d\: | cut -f1 -d,`
-#  echo soilTemperature.min = $SLMIN
-#  echo soilTemperature.max = $SLMAX
-#  echo soilTemperature.avg = $SLAVG
-   read line  # should be '}'
+      read SOILTMP # should be like :
+if [ "$DEBUG" = "true" ] ; then
+echo SOILTEMP $SOILTMP
+fi
+      # "min":24.3,"max":24.7,"avg":24.5
+      SLMIN=`echo $SOILTMP | cut -f2 -d\: | cut -f1 -d,`
+      SLMAX=`echo $SOILTMP | cut -f3 -d\: | cut -f1 -d,`
+      SLAVG=`echo $SOILTMP | cut -f4 -d\: | cut -f1 -d,`
 
-   # "deltaT"
-   read line  # should be ',"deltaT":'
+      read line  # should be '}'
+      # "deltaT"
+      read line  # should be ',"deltaT":'
+   else
+      SLMIN=""
+      SLMAX=""
+      SLAVG=""
+   fi
+if [ "$DEBUG" = "true" ] ; then
+   echo soilTemperature.min = $SLMIN
+   echo soilTemperature.max = $SLMAX
+   echo soilTemperature.avg = $SLAVG
+fi
    read line  # should be '{'
 
    read DELTAT # should be like :
@@ -126,13 +150,36 @@ extract_entry () {
    DLMIN=`echo $DELTAT | cut -f2 -d\: | cut -f1 -d,`
    DLMAX=`echo $DELTAT | cut -f3 -d\: | cut -f1 -d,`
    DLAVG=`echo $DELTAT | cut -f4 -d\: | cut -f1 -d,`
-#  echo deltaT.min = $DLMIN
-#  echo deltaT.max = $DLMAX
-#  echo deltaT.avg = $DLAVG
+if [ "$DEBUG" = "true" ] ; then
+   echo deltaT.min = $DLMIN
+   echo deltaT.max = $DLMAX
+   echo deltaT.avg = $DLAVG
+fi
    read line  # should be '}'
+
+   # "wetBulb"
+   read line  # should be ',"wetBulb":'
+if [ "$DEBUG" = "true" ] ; then
+echo WETBULD $line
+fi
+   read line  # should be '{'
+
+   read WETBULB # should be like :
+   # "max":15.7,"avg":15
+   WBMAX=`echo $WETBULB | cut -f2 -d\: | cut -f1 -d,`
+   WBAVG=`echo $WETBULB | cut -f3 -d\:`
+if [ "$DEBUG" = "true" ] ; then
+   echo wetBulb.max = $WBMAX
+   echo wetBulb.avg = $WBAVG
+fi
+   read line  # should be '}'
+
 
    # "wind"
    read line  # should be ',"wind":['
+if [ "$DEBUG" = "true" ] ; then
+  echo WIND LINE $line
+fi
    read line  # should be '{'
 
    # ,"wind":[
@@ -159,21 +206,27 @@ extract_entry () {
    read line
       # "height":3,"avg":
    WHEIGHT=`echo $line | cut -f2 -d\: | cut -f1 -d,`
-#  echo wind.height = $WHEIGHT
+if [ "$DEBUG" = "true" ] ; then
+   echo wind.height = $WHEIGHT
+fi
 
    read line  # should be '{'
    read line
    # "speed":4.24,"direction":
    WSPAVG=`echo $line | cut -f2 -d\: | cut -f1 -d,`
-#  echo wind.avg.speed = $WSPAVG
+if [ "$DEBUG" = "true" ] ; then
+   echo wind.avg.speed = $WSPAVG
+fi
 
    read line  # should be '{'
    read line
    # "degrees":158,"compassPoint":"SSE"
    WSPADIR=`echo $line | cut -f2 -d\: | cut -f1 -d,`
    WSPACMP=`echo $line | cut -f3 -d\:`
-#  echo wind.max.dir = $WSPDIR
-#  echo wind.max.compass = $WSPCMP
+if [ "$DEBUG" = "true" ] ; then
+   echo wind.max.dir = $WSPDIR
+   echo wind.max.compass = $WSPCMP
+fi
    read line  # should be '}'
    read line  # should be empty
    read line  # should be '}'
@@ -183,14 +236,20 @@ extract_entry () {
    read line
    # "speed":10.84,"time":"2016-11-30T23:02:00Z","direction":
    WSPMAX=`echo $line | cut -f2 -d\: | cut -f1 -d,`
-#  echo wind.max.speed = $WSPMAX
+if [ "$DEBUG" = "true" ] ; then
+   echo wind.max.speed = $WSPMAX
+fi
    read line  # should be '{'
    read line
    # "degrees":139,"compassPoint":"SE"
    WSPMDIR=`echo $line | cut -f2 -d\: | cut -f1 -d,`
-#  echo wind.max.dir = $WSPMDIR
+if [ "$DEBUG" = "true" ] ; then
+   echo wind.max.dir = $WSPMDIR
+fi
    WSPMCMP=`echo $line | cut -f3 -d\:`
-#  echo wind.max.compass = $WSPMCMP
+if [ "$DEBUG" = "true" ] ; then
+   echo wind.max.compass = $WSPMCMP
+fi
    read line  # should be '}'
 
    read line  # should be empty
@@ -249,8 +308,10 @@ get_data() {
 
    STARTTIME=`date --date=${2}-${3}-${4} +%F`T01:00Z
    ENDTIME=`date --date=${2}-${3}-${4}+1day +%F`T00:00Z
-#  echo start $STARTTIME
-#  echo end   $ENDTIME
+if [ "$DEBUG" = "true" ] ; then
+   echo start $STARTTIME
+   echo end   $ENDTIME
+fi
 
    OUTFILE="dpird_${1}_daily_${2}${3}${4}.csv"
 
@@ -272,15 +333,21 @@ get_data() {
 # URL="${URL},wind,windAvgSpeed,windMaxSpeed"
 # CURLCMD="curl -X GET ${URL} -H accept:application/json -H API-KEY:${APIKEY} -s -o ${FILE}"
 
-#  echo "Curling ${STARTTIME} - ${ENDTIME}"
+if [ "$DEBUG" = "true" ] ; then
+   echo "Curling ${STARTTIME} - ${ENDTIME}"
+fi
 
    CURLCMD="curl -X GET ${HOST}v2/weather/stations/${STATION}/summaries/hourly?startDateTime=${STARTTIME}&endDateTime=${ENDTIME}&offset=0&limit=25&select=airTemperature,airTemperatureAvg,airTemperatureMax,airTemperatureMin,deltaTAvg,deltaTMax,deltaTMin,dewPointAvg,dewPointMax,dewPointMin,etoShortCrop,etoTallCrop,rainfall,relativeHumidity,relativeHumidityAvg,relativeHumidityMax,relativeHumidityMin,soilTemperatureAvg,soilTemperatureMax,soilTemperatureMin,solarExposure,wetBulbAvg,wetBulbMax,wind,windAvgSpeed,windMaxSpeed -H accept:application/json -H API-KEY:${APIKEY} -s -o ${FILE}"
-#  echo ${CURLCMD}
+if [ "$DEBUG" = "true" ] ; then
+   echo ${CURLCMD}
+fi
    ${CURLCMD}
 
    if [ -f ${FILE} ] ; then
 
-#    echo "All curly"
+if [ "$DEBUG" = "true" ] ; then
+     echo "All curly"
+fi
 
      # The header should be like (only 1 line though) :
      # station_id,record_datetime,air_temp_min,air_temp_max,air_temp_avg_degC,
@@ -296,7 +363,9 @@ get_data() {
 
      sed -e 's/{/\n{\n/g' -e 's/}/\n}\n/g' < ${FILE} | extract_data
 
+if [ "$DEBUG" != "true" ] ; then
      /bin/rm ${FILE}
+fi
    fi
 }
 #------------------------------------------------------------------------------#
@@ -317,7 +386,7 @@ get_data() {
 # else
 #    /bin/rm ${OUTFILE}
   fi
-  exit 0
+# exit 0
 
 . ./common/finish.sh
